fixed bugs

+++ Services/ManageOccupants.cs
@@ -1,19 +1,19 @@
 ﻿using DtnParkingSystem.Interface;
 using DtnParkingSystem.Models;
-using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace DtnParkingSystem.Services
 {
-	public class ManageOccupants : IManageOccupants
+    public class ManageOccupants : IManageOccupants
     {
         private readonly IOccupantsDAO _occupantsDAO;
-
-        public ManageOccupants(IOccupantsDAO occupantsDAO)
+        private readonly IParkingSpaceDAO _parkingSpaceDAO;
+        public ManageOccupants(IOccupantsDAO occupantsDAO, IParkingSpaceDAO parkingSpaceDAO)
         {
             _occupantsDAO = occupantsDAO;
+            _parkingSpaceDAO = parkingSpaceDAO;
         }
 
-        public async Task<string>Register(string fullName, string contactNumber, string plateNumber, string vehicleType)
+        public async Task<string> Register(string fullName, string contactNumber, string plateNumber, string vehicleType)
         {
             try
             {
@@ -36,11 +36,11 @@ namespace DtnParkingSystem.Services
                     return (validate);
                 }
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 return ("Error:" + ex);
             }
-  
+
         }
 
 
@@ -77,25 +77,57 @@ namespace DtnParkingSystem.Services
                 {
                     return ("Please fill all fields");
                 }
-            }catch(Exception ex)
+            }
+            catch (Exception ex)
             {
                 return ("Error: " + ex);
             }
-        
+
         }
         public string Delete(string fullName)
         {
             try
             {
-                _occupantsDAO.DeleteUser(fullName);
-                return ("Success");
+                var validateResult = ValidateDelete(fullName).Result;
+                if (validateResult == "Success")
+                {
+                    _occupantsDAO.DeleteUser(fullName);
+                    return ("Success");
+                }
+                else
+                {
+                    return (validateResult);
+                }
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 return ("Error: " + ex);
             }
         }
 
+        public async Task<string> ValidateDelete(string fullName)
+        {
+            var occupantsinAllFloors = await _parkingSpaceDAO.GetDocuments<ParkingSpaces>(default);
+            var listOccupants = new List<string> { };
+            foreach (var items in occupantsinAllFloors)
+            {
+                var floorOccupants = await _parkingSpaceDAO.GetSubCollection<ParkingSpaces>(items, default);
+
+                foreach (var occupants in floorOccupants)
+                {
+                    listOccupants.Add(occupants.Occupant);
+                }
+            }
+
+            foreach (var items in listOccupants)
+            {
+                if (fullName == items)
+                {
+                    return ("Please free up the slot this user has occupied first");
+                }
+            }
+            return ("Success");
+        }
         public async Task<string> Validate(string fullName, string contactNumber, string plateNumber, string vehicleType)
         {
             if (!string.IsNullOrEmpty(fullName) && !string.IsNullOrEmpty(contactNumber) && !string.IsNullOrEmpty(vehicleType))
@@ -114,7 +146,7 @@ namespace DtnParkingSystem.Services
                     }
 
                 }
-            
+
                 return ("Success");
             }
             else
